requirements.txt
Flask
Flask-RESTful
Flasgger
gunicorn
firebase-admin
werkzeug
python-dotenv
flask-cors